changed environment property name

+++ development/ui/dashboard_ui/education_usecase/client-side/src/app/home-page/home-page.component.ts
@@ -11,49 +11,60 @@ import { AppServiceComponent } from '../app.service';
 })
 export class HomePageComponent implements OnInit {
   adminUrl;
+  adminDashUrl
   constructor(public keycloakService: KeycloakSecurityService, public router: Router, public service: AppServiceComponent) {
     service.logoutOnTokenExpire();
   }
 
   ngOnInit(): void {
     this.adminUrl = environment.adminUrl;
+
     if (localStorage.getItem('roleName') != 'admin') {
       this.router.navigate(['/home']);
     }
-    if (this.keycloakService.kc.tokenParsed.realm_access) {
-      if (this.keycloakService.kc.tokenParsed.realm_access.roles.includes('admin')) {
-        localStorage.setItem('roleName', 'admin');
-        this.router.navigate(['/home']);
-      } else if (this.keycloakService.kc.tokenParsed.realm_access.roles.includes('report_viewer')) {
-        localStorage.setItem('roleName', 'report_viewer');
-        this.router.navigate(['/dashboard/infrastructure-dashboard']);
-      } else {
-        if (!this.keycloakService.kc.tokenParsed.realm_access.roles.includes('report_viewer')
-          || !this.keycloakService.kc.tokenParsed.realm_access.roles.includes('admin')) {
-          alert("Unauthorised user, Only admin and viewer can login")
-          let options = {
-            redirectUri: environment.appUrl
+    if (environment.auth_api === 'cqube') {
+      if (this.keycloakService.kc.tokenParsed.realm_access) {
+        if (this.keycloakService.kc.tokenParsed.realm_access.roles.includes('admin')) {
+          localStorage.setItem('roleName', 'admin');
+          this.router.navigate(['/home']);
+        } else if (this.keycloakService.kc.tokenParsed.realm_access.roles.includes('report_viewer')) {
+          localStorage.setItem('roleName', 'report_viewer');
+          this.router.navigate(['/dashboard/infrastructure-dashboard']);
+        } else {
+          if (!this.keycloakService.kc.tokenParsed.realm_access.roles.includes('report_viewer')
+            || !this.keycloakService.kc.tokenParsed.realm_access.roles.includes('admin')) {
+            alert("Unauthorised user, Only admin and viewer can login")
+            let options = {
+              redirectUri: environment.appUrl
+            }
+            this.keycloakService.kc.logout(options);
           }
-          this.keycloakService.kc.logout(options);
         }
+      } else {
+        alert("Please assign role to user");
+        let options = {
+          redirectUri: environment.appUrl
+        }
+        sessionStorage.clear();
+        this.keycloakService.kc.clearToken();
+        this.keycloakService.kc.logout(options);
       }
-    } else {
-      alert("Please assign role to user");
+    }
+  }
+
+  logout() {
+    if (environment.auth_api === 'cqube') {
+      localStorage.clear();
       let options = {
         redirectUri: environment.appUrl
       }
-      sessionStorage.clear();
       this.keycloakService.kc.clearToken();
       this.keycloakService.kc.logout(options);
+    } else {
+      localStorage.clear();
+      this.router.navigate(['signin'])
     }
   }
-  logout() {
-    localStorage.clear();
-    let options = {
-      redirectUri: environment.appUrl
-    }
-    this.keycloakService.kc.clearToken();
-    this.keycloakService.kc.logout(options);
-  }
+
 
 }
